+++ formats/html.php
@@ -28,7 +28,7 @@ class HTMLFormat implements Format {
 <script>
 // this is a special type of icon. You can remove this class if you don't need it
 L.PopupIcon = L.Icon.extend({
-    initialize: function( text, options ){
+    initialize: function( text, options ) {
         L.Icon.prototype.initialize.call(this, options);
         this._text = text;
     },
@@ -42,7 +42,10 @@ L.PopupIcon = L.Icon.extend({
         div.style.position = 'absolute';
         div.style.width = width + 'px';
         div.style.bottom = '-3px';
+        div.style.pointerEvents = 'none';
         div.style.left = (-width / 2) + 'px';
+        div.style.margin = div.style.padding = '0';
+        pdiv.style.margin = pdiv.style.padding = '0';
 
         var contentDiv = document.createElement('div');
         contentDiv.innerHTML = this._text;
@@ -54,24 +57,33 @@ L.PopupIcon = L.Icon.extend({
         contentDiv.style.borderRadius = '5px';
         contentDiv.style.margin = '0 auto';
         contentDiv.style.display = 'table';
+        contentDiv.style.pointerEvents = 'auto';
+
+        var stop = L.DomEvent.stopPropagation;
+        L.DomEvent
+            .on(contentDiv, 'click', stop)
+            .on(contentDiv, 'mousedown', stop)
+            .on(contentDiv, 'dblclick', stop);
 
         var tipcDiv = document.createElement('div');
         tipcDiv.className = 'leaflet-popup-tip-container';
         tipcDiv.style.width = '20px';
         tipcDiv.style.height = '11px';
+        tipcDiv.style.padding = '0';
+        tipcDiv.style.margin = '0 auto';
         var tipDiv = document.createElement('div');
         tipDiv.className = 'leaflet-popup-tip';
         tipDiv.style.width = tipDiv.style.height = '8px';
         tipDiv.style.marginTop = '-5px';
         tipDiv.style.boxShadow = 'none';
         tipcDiv.appendChild(tipDiv);
-        
+
         div.appendChild(contentDiv);
         div.appendChild(tipcDiv);
         pdiv.appendChild(div);
         return pdiv;
     },
-    
+
     createShadow: function () {
         return null;
     }